reverted to earlier App.svelte version
