+++ client/jsconfig.json
@@ -28,5 +28,5 @@
    * Use global.d.ts instead of compilerOptions.types
    * to avoid limiting type declarations.
    */
-  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.svelte", "src/components/PromptInput copy_svelte.txt"]
+  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.svelte", "src/components/PromptInput copy_svelte.txt", "src/components/PreviewWindow.svelte.bak", "src/App.svelte.new", "src/components/PreviewWindow.svelte.bak2"]
 }
