+++ client/src/components/PreviewWindow.svelte.err.bak
@@ -1,0 +1,71 @@
+<script>
+  import PromptInput from './PromptInput.svelte';
+  import PreviewWindow from './PreviewWindow.svelte.bak2';
+</script>
+
+<main>
+  <header>
+    <h1>AetherPress</h1>
+  </header>
+
+  <div class="container">
+    <div class="input-panel">
+      <PromptInput />
+      <button class="generate-button">Generate</button>
+    </div>
+    <div class="preview-panel">
+      <PreviewWindow />
+    </div>
+  </div>
+</main>
+
+<style>
+  :root {
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
+      Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  }
+
+  main {
+    text-align: center;
+    padding: 1em;
+    max-width: 800px;
+    margin: 0 auto;
+  }
+
+  header {
+    margin-bottom: 1.5rem;
+  }
+
+  .container {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+  }
+
+  .input-panel {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+
+  .generate-button {
+    padding: 0.75rem 1.5rem;
+    font-size: 1rem;
+    border: none;
+    border-radius: 6px;
+    background-color: #007acc;
+    color: white;
+    cursor: pointer;
+  }
+
+  .generate-button:hover {
+    background-color: #005fa3;
+  }
+
+  .preview-panel {
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    padding: 1rem;
+    min-height: 300px;
+  }
+</style>
